fix: Update release action
