+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -10,6 +10,11 @@
 ## Release Category
 
 <!-- Documentation, Infrastructure, Tokens -->
+<!-- For packiage specific changes, add package name before category: Web Tokens, Web Documentation, etc -->
+
+### Release Note
+Optional release note message. Add `⚠ BREACKING CHANGES:` before message if it's a breaking change.
+Changelog and release summaries will contain a pull request title. This section will add additional notes under that title. This section is not a summary, but something extra to point out in release notes. An example might be calling out breaking changes in a labs component or minor visual changes that need visual regression updates. Remove this section if no additional release notes are required.
 
 ---
